Rename and refactor test files for chat model organization
Moved test files to a more specific package structure for clarity. Removed unused imports and a redundant test to improve code cleanliness and maintainability.
